Moved static url replacement to new mixin in lms_xblock.runtime
Had to move the previous change out of common.lib due to import issues
encountered when building assets.  common.lib shouldn't depend on djangoapps.
