+++ cms/djangoapps/contentstore/views/preview.py
@@ -20,6 +20,7 @@ from cms.lib.xblock.field_data import CmsFieldData
 from contentstore.utils import get_visibility_partition_info
 from contentstore.views.access import get_user_role
 from edxmako.shortcuts import render_to_string
+from lms.djangoapps.lms_xblock.runtime import StaticUrlsRuntimeMixin
 from lms.djangoapps.lms_xblock.field_data import LmsFieldData
 from openedx.core.lib.license import wrap_with_license
 from openedx.core.lib.xblock_utils import (
@@ -91,7 +92,7 @@ def preview_handler(request, usage_key_string, handler, suffix=''):
     return webob_to_django_response(resp)
 
 
-class PreviewModuleSystem(ModuleSystem):  # pylint: disable=abstract-method
+class PreviewModuleSystem(StaticUrlsRuntimeMixin, ModuleSystem):  # pylint: disable=abstract-method
     """
     An XModule ModuleSystem for use in Studio previews
     """
